requirements.txt
langchain
gitpython
python-dotenv